跑通基本逻辑

+++ internal/model/params.go
@@ -10,7 +10,7 @@ import (
 type Params struct {
 	Dim        int64   `json:"dim"`
 	MultipleOf int     `json:"multiple_of"`
-	Heads      int     `json:"n_heads"`
+	Heads      int64   `json:"n_heads"`
 	Layers     int     `json:"n_layers"`
 	Eps        float32 `json:"norm_eps"`
 	Vocabs     int     `json:"vocab_size"`
